+++ docs/package.json
@@ -6,10 +6,11 @@
     "src"
   ],
   "scripts": {
-    "check-links": "tsx validate-docs-links.ts"
+    "check-links": "tsx src/validate-docs-links.ts"
   },
   "devDependencies": {
     "@types/github-slugger": "^1.3.0",
+    "@types/node": "^22.1.0",
     "@vercel/ncc": "0.34.0",
     "tsx": "^4.7.2",
     "typescript": "5.1.6"
